requirements.txt
numpy
python-mnist
tensorflow
torch
torchvision
futhark-server
futhark-data
